requirements.txt
Django>5.1
nanodjango
Pillow 
python-decouple
django-filter
django-htmx